Fixed update script.
Wordpress changed their readme.html file so that the full version number is no
longer mentioned (it now says: Version 4.7 for all versions 4.7.*). This is no
bueno for our script, as we want to prevent downloading the same files all the
time so we need to know the version. Since Wordpress provides no other way to
get the version, I opted instead to use the more robust SHA1 hash of the
latest.tar.gz file to determine whether we have the latest version of Wordpress.

This saves us from having to slice and dice strings using awk and perl, and
works around the fact that the Wordpress people removed the full version number
from the readme.

I also fixed a little issue with the verbose checking.